requirements.txt
rasterio~=1.2.3
matplotlib~=3.4.2
numpy~=1.20.3
pillow~=8.2.0
h5py~=3.2.1
geopandas~=0.9.0
shapely~=1.7.1
dbfread~=2.0.7
dbf~=0.99.1
sphinx